[jaxrs-cxf-cdi] generate fluent interface "allOf"
When using allOf in the model, it is convenient to generated overridden
fluent setters, in order to return the super type instead of the base
type.
Otherwise, one has to use the setters/getters, and this makes the users
code more complex and harder to read.

I added values to the parentVars model field, using the same technique
as used in the csharp model.