[go][go-experimental] Fixes problems with conflicting parameters
OpenAPI allows parameters to bear the same name if they appear in
different settings inside the same operation.

The code generated by the go-experimental generator, though, could
sometimes fail to compile due to parameters with the same name. Both the
field in the api*Request struct and the setter method names can clash if
there are multiple parameteres with the same name.

This change fixes that by namespacing all the param names in the
generated code with their origin.