Treat classes from java.lang as reserved words
Classes in the package `java.lang` are available without explicitly
importing them.

This leads to naming collisions in the generated code if there is a
schema named like a class in `java.lang`.

For example a schema named `override` would clash with
`java.lang.Override`, an annotation used in virtually every template for
the Java generators.

As the least intrusive method (with regards to existing templates), we
handle classes from `java.lang` as reserved words, so that they will be
renamed accordingly.