[rust-server] Add support for Bearer auth
As far as I can tell, the bulk of the work to support bearer auth was already implemented to support OAuth, but never explicitly turned on. I've added a new sample which appears to be working correcltly, but I may have unwittingly overlooked something.

Changes:
- Added bearer auth to the security features for rust server
- Supplemented the basic auth condition in the context template to handled basic auth and bearer auth separately.
- Repurpose an exising sample to confirm the code generation works as expected.

The new sample builds without problems, so the only possibly issue is that the underlying `swagger` crate didn't implement this. However, there's already bearer support in the client and it's also used in the OAuth support in the server, so I don;t see why it wouldn't work.