rust: better support for binary body parameters
The Rust-based reqwest client does not correctly handle a binary body
parameter today.  It creates a PathBuf parameter and attempts to pass
that to the RequestBuilder json() method, which sends the path itself
as the body of the request.

Instead, it would help to parameterise operations that accept a binary
body so that they may accept any object that implements the Into<Body>
trait.  This would allow the caller to pass a binary body in several
forms, such as a slice of bytes or a Stream that reads from some other
source.