Use Kotlin nullable operator for nullable properties.
The previous check of isReadOnly was not applicable to whether a value should be null or not. Also removed the default null value if the value is not nullable.

Also removed escaping of default values (this would cause any string value to be written to the file as val test = &quot;value&quot; instead of val test = "value"