[required_vars_changes]
Problem:
The following PR removed the logic for adding the default value
for required_vars in composed model's __init__ and _from_openapi_data
method due to which it is failing in allOf model class stating missing
required positional argument.

https://github.com/OpenAPITools/openapi-generator/pull/8802

Now it will become users responsibility to add the required_vars as it will
become mandatory input in the payload.

Since in our application end users may not be knowing these required_vars
all the time, the SDK failure create an issue for us.

Solution:
The proposed solution is to add a configuration variable: setRequiredVars.
If it is set to true, then the older logic will be used. The older logic
was using to set the default value of required_vars in __init__ and
_from_openapi_data as positional arguement.