Remove dummy variable declaration for context.Context from generated go code
Originally, this dummy declaration was needed because there was the
possibility of the generated code doesn't use `context.Context` and then
if it imported that package, go compiler complains that and makes an error.
ref: https://github.com/OpenAPITools/openapi-generator/blob/3ed1aa8e79687bed63dfa064de27317273080471/modules/swagger-codegen/src/main/resources/go/api.mustache#L30

However, now this dummy placement is no longer needed because the
generated code always uses `context.Context`.

Signed-off-by: moznion <moznion@mail.moznion.net>