[jaxrs-cxf-cdi] mark deprecated api methods (#11755)
In the generated java code, it is usefull to have the generated java api
use the @Deprecated annotation. This enable to leverage this annotation
at runtime to trigger specific logging for instance.

In the generated interface, use the @Deprecated annotation.
Also, in the jax-rs implementation that links the jax-rs api with the
cdi bean implementing the interface, use the
@SuppressWarnings("deprecation") annotation. This way the deprecation
warning is not shown in generated code.

One can use the interface as before: the java compiler can then tell
which implementations do implement a deprecated api, using a warning.
If the implementation itself also sets the @Deprecated annotation, then
there is no warning, unless the implementation is called at another
place in the code.