Fix nullable models embedded in allOf
It is possible, though contrived, for a property to have both isAnyType
and isModel set. In this case, when the outer type is nullable, the Go
template expects a Nullable* type to be provided by the generator, but
it is skipped over instead. This change aligns the generator with the
template's expectations.