Fix oneOf discriminator mapping in allOf
There are two ways that discriminators can be used: with oneOf or anyOf,
explicitly, alongside the definition of their respective schemas; or
with allOf, in which case a discriminator defined in a parent schema may
be implicitly inherited by any other schema that includes it.

The code generator correctly handles both of those cases, except when an
explicit ("closed") oneOf is used directly inside an unrelated allOf
composite schema. In this situation, the allOf leaks into the
discriminator mappings for the oneOf schema.

This change checks whether a schema for which a discriminator is being
created is a oneOf/anyOf and skips the "open" allOf check that follows
if so.