Flatten inline schemas in requestBody components
For request bodies that are completely inlined into a path's operations,
the inline model resolver correctly generates both a schema component
and a requestBody component and adds $refs.

However, when an operation already contains an $ref to a requestBody
that itself contains an inline schema, that schema was not being
flattened to a separate component. This change adds a separate method to
do such flattening and add an $ref if needed.