[java-micronaut] make parent with a discriminator abstract and sealed
Parent types with a discriminator which aren't part of the mapped types
could be abstract, since they can't be instantiated themself. This is
however only the case if all possible types are part of the mapping.

By adding a configuration option this behavior becomes configurable in
order to let the developer decide if the parent type must be abstract
and sealed.